% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\docType{data}
\name{area}
\alias{area}
\title{Area data table}
\value{
area data.frame with all available lake areas

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
area \tab numeric \tab lake area in m^2\cr
}
}
\examples{
head(area)

}

