% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\docType{data}
\name{canopy}
\alias{canopy}
\title{Canopy heights}
\format{A data frame with three columns: \code{site_id}, \code{canopy_m},
  and \code{source}.}
\usage{
canopy
}
\description{
Canopy heights in meters for lakes indexed by NHD ID.
}
\keyword{datasets}

