% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area.R
\name{get_area}
\alias{get_area}
\title{Get lake surface area}
\usage{
get_area(id)
}
\arguments{
\item{id}{The character ID for the requested data}
}
\value{
Lake surface areas in meters^2. NA if no value available
}
\description{
Returns the surface area for a lake with given ID
}
\details{
Looks for given id and returns values if availalbe. id
is coerced to character regardless of input type.
}
\examples{
#Multiple values
get_area('nhd_1097324')

#this should return NA
get_area('asdf')

}
