% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kd_year.R
\name{get_kd_year}
\alias{get_kd_year}
\title{Get light attenuation coefficient for a lake through time}
\usage{
get_kd_year(id, years, src = "in-situ")
}
\arguments{
\item{id}{The character ID for the requested data}

\item{years}{Integer vector of years for which you want Kd values}

\item{src}{The data source, currently one of \code{c('in-situ', 'satellite')}. In-situ observations are default}
}
\value{
data.frmae of Kd values with the same number of rows as \code{length(years)}. NA filled when missing.
}
\description{
Returns the annual average light attenuation coefficient for a lake with the given ID
}
\details{
TODO
}
\examples{
#Get and plot Kds for a lake over time
years = 1970:2012
kds   = get_kd_year('WBIC_1881900', years)
plot(secchi_m~year, kds, type='o')


}
\author{
Luke Winslow, Jordan Read
}

