% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latlon.R
\name{get_latlon}
\alias{get_latlon}
\title{Get latitude and longitude for a given lake}
\usage{
get_latlon(id)
}
\arguments{
\item{id}{The character ID for the requested data}
}
\value{
Lat/lon on the WGS84 datum as two-element vector in order lat, long. If not ID match found, 
 return NA values.
}
\description{
Get the center lat/lon of a lake with given ID
}
\details{
Based on the exported locations data.frame. Mostly based on NHD polygon centroid lat/lons.
}
\examples{
get_latlon('nhd_1097324')


}

