% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_turbidity_avg.R
\name{get_turbidity_avg}
\alias{get_turbidity_avg}
\title{Get lake average turbidity}
\usage{
get_turbidity_avg(ids, src = "in-situ")
}
\arguments{
\item{ids}{The character ID for the requested kd data. Can be a character vector.}

\item{src}{The data source, currently one of \code{c('in-situ', 'satellite')}. In-situ observations are default}
}
\description{
Get lake average turbidity
}
\author{
Luke Winslow
}
