% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zmax.R
\name{get_zmax}
\alias{get_zmax}
\title{Get max depth for a given lake}
\usage{
get_zmax(id)
}
\arguments{
\item{id}{The character ID for the requested data}
}
\value{
Max observed depth in meters
}
\description{
Get the max depth for a given lake id
}
\details{
TODO
}
\examples{
get_zmax('nhd_1097324')


}
