\name{lakeattributes-package}
\alias{lakeattributes-package}
\alias{lakeattributes}
\docType{package}
\title{
\packageTitle{lakeattributes}
}
\description{
\packageDescription{lakeattributes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lakeattributes}
\packageIndices{lakeattributes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lakeattributes}

Maintainer: \packageMaintainer{lakeattributes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
