% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location.R
\docType{data}
\name{location}
\alias{location}
\title{Location data table}
\value{
Location data.frame with all available lake locations. 
Usually defined as lake polygon centroid.

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
lon \tab numeric \tab Lake Longitude\cr
lat \tab numeric \tab Lake Latitude\cr
}
}
\examples{
head(area)

}

