% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhd2wbic.R
\docType{data}
\name{nhd2wbic}
\alias{nhd2wbic}
\title{NHD ID to WBIC (Wisconsin) ID table}
\format{nhd2wbic data.frame site_id (NHD) and WBIC columns

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
WBIC \tab numeric \tab Wisconsin Water Body Identifier code \cr
}}
\description{
NHD ID to WBIC (Wisconsin) ID table
}
\examples{
data(nhd2wbic)
head(nhd2wbic)

}
