% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption_helpers.R
\name{read_aes}
\alias{read_aes}
\title{read encypted data frame from file}
\usage{
read_aes(filename, key)
}
\arguments{
\item{filename}{Location to save or load encrypted binary file}

\item{key}{A secret key as a character string value. Must be <= 32 characters in length}
}
\description{
WARNING: Do not use for security critical applications. I am not a security expert. 
This is for low-security applications only.

This unencrypts and reads a file written with \code{\link{write_aes}}. Uses defaults for
\code{\link{write.csv}}. This could cause issues with character columns with odd characters.
}
\seealso{
\code{\link{write_aes}}.
}

