% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secchi.R
\docType{data}
\name{secchi}
\alias{secchi}
\title{Secchi data table}
\value{
secchi data.frame with all available observed secchi data.

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
year \tab numeric \tab Year of observation\cr
date \tab character \tab Date of observation (May 15th for secchi averages)\cr
secchi_m \tab numeric \tab Observed or estimated Secchi depth in meters\cr
source \tab character \tab Source (satellite or in-situ)\cr
}
}
\examples{
head(secchi)

}

