% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtemp.R
\docType{data}
\name{wtemp}
\alias{wtemp}
\title{Water Temperature data table}
\value{
wtemp data.frame with all available observed water temp data.

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
year \tab numeric \tab Year of observation\cr
date \tab character \tab Date of observation\cr
wtr \tab numeric \tab Observed water temp in deg C\cr
source \tab character \tab Source (currently only in-situ)\cr
}
}
\examples{
head(wtemp)

}

