% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmax.R
\docType{data}
\name{zmax}
\alias{zmax}
\title{zmax data table}
\value{
zmax data.frame with all available lake max depth values

\tabular{lll}{
site_id \tab character \tab Unique identifier of site\cr
zmax_m \tab numeric \tab lake max depth in meters\cr
}
}
\description{
zmax data table
}
\examples{
head(zmax)

}
