% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\docType{data}
\name{canopy}
\alias{canopy}
\title{Lake surrounding canopy height}
\format{An object of class \code{data.frame} with 17437 rows and 3 columns.}
\usage{
canopy
}
\value{
canopy data.frame with all available observed heights

\tabular{lll}{
site_id \tab character \tab Unique identifier of site\cr
canopy_m \tab numeric \tab Estimated canopy height in meters\cr
source \tab character \tab Source of data (NLCD, satellite, etc)\cr
}
}
\description{
Canopy height in meters derived from a variety of sourced
linked to the NHD PermIDs.
}
\examples{
data(canopy)
head(canopy)

}
\keyword{datasets}
