% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bathy.R
\name{get_bathy}
\alias{get_bathy}
\title{Get Hypsometry for a given lake}
\usage{
get_bathy(site_id)
}
\arguments{
\item{site_id}{The character ID for the requested data}
}
\value{
Data frame with columns \code{depth} and \code{area}
}
\description{
Returns the hypsometry profile for a lake with the given ID
}
\examples{
get_bathy('nhd_1097324')


}

