% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kd_best.R
\name{get_kd_best}
\alias{get_kd_best}
\title{Return best available secchi/Kd}
\usage{
get_kd_best(id, years, datasource = NA)
}
\arguments{
\item{id}{The character ID for the requested data}

\item{years}{Integer vector of years for which you want Kd values}

\item{datasource}{NA (for no filter) or vector of string sources to filter. Get options in dataset 
using \code{unique(secchi$datasource)}.}
}
\value{
A data frame with columns 'year', 'secchi_avg', and 'source'.
}
\description{
Return best available secchi/Kd
}
\details{
This implements a hierarchical selection routine. The hierarchy is:
\enumerate{
\item Mean in-situ secchi data for each year available
\item Mean satellite secchi data for each year availble
\item Overall mean in-situ secchi value for the lake
\item Overall mean satellite secchi value for the lake
\item Population mean of in-situ data (currently 2.46566 meters)
}

This guarantees you will always get a secchi value for a lake. This 
does not mean that lake has observed secchi data, as 'population mean' 
is the catch-all.
}
\examples{

#get the best available secchi depths for Sparkling lake
get_kd_best('WBIC_1881900', 1979:2012)


}
