% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtemp.R
\docType{data}
\name{wtemp}
\alias{wtemp}
\title{Water Temperature data table}
\format{An object of class \code{data.frame} with 1578066 rows and 6 columns.}
\usage{
wtemp
}
\value{
wtemp data.frame with all available observed water temp data.

\tabular{lll}{
Name \tab Type \tab Description\cr
site_id \tab character \tab Unique identifier of site\cr
year \tab numeric \tab Year of observation\cr
date \tab character \tab Date of observation\cr
wtr \tab numeric \tab Observed water temp in deg C\cr
source \tab character \tab Source (currently only in-situ)\cr
}
}
\description{
Water temperature observations collected from a variety of
sourced. Primarily profile data, indexed by NHD PermID.
}
\examples{
data(wtemp)
head(wtemp)

}
\keyword{datasets}
